+++ ogre2/include/ignition/rendering/ogre2/Ogre2AxisVisual.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015 Open Source Robotics Foundation
+ * Copyright (C) 2020 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  *
  */
-#ifndef IGNITION_RENDERING_AXISVISUAL_HH_
-#define IGNITION_RENDERING_AXISVISUAL_HH_
+#ifndef IGNITION_RENDERING_OGRE2_OGRE2AXISVISUAL_HH_
+#define IGNITION_RENDERING_OGRE2_OGRE2AXISVISUAL_HH_
 
-#include "ignition/rendering/config.hh"
-#include "ignition/rendering/CompositeVisual.hh"
+#include "ignition/rendering/base/BaseAxisVisual.hh"
+#include "ignition/rendering/ogre2/Ogre2Visual.hh"
 
 namespace ignition
 {
@@ -26,17 +26,18 @@ namespace ignition
   {
     inline namespace IGNITION_RENDERING_VERSION_NAMESPACE {
     //
-    /// \class AxisVisual AxisVisual.hh ignition/rendering/AxisVisual.hh
-    /// \brief Represents a axis composite visual
-    class IGNITION_RENDERING_VISIBLE AxisVisual :
-      public virtual CompositeVisual
+    /// \brief Ogre2.x implementation of the axis visual class
+    class IGNITION_RENDERING_OGRE2_VISIBLE Ogre2AxisVisual :
+      public BaseAxisVisual<Ogre2Visual>
     {
-      /// \brief Deconstructor
-      public: virtual ~AxisVisual() { }
+      /// \brief Constructor
+      protected: Ogre2AxisVisual();
 
-      /// \brief set true to show the axis heads, false otherwise
-      /// \param[in] _b true to show the axis heads, false otherwise
-      public: virtual void ShowAxisHead(bool _b) = 0;
+      /// \brief Destructor
+      public: virtual ~Ogre2AxisVisual();
+
+      /// \brief Only scene can instantiate an axis visual
+      private: friend class Ogre2Scene;
     };
     }
   }